+++ app/src/main/java/com/samnagaraj/simplechess/ChessBoardView.kt
@@ -1,16 +1,18 @@
-package com.example.chess_app
+package com.samnagaraj.simplechess
 
 import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.Paint
 import android.graphics.RectF
-import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
-import android.widget.Toast
 import androidx.core.content.ContextCompat
+import com.samnagaraj.simplechess.model.ChessGame
+import com.samnagaraj.simplechess.model.Piece
+import com.samnagaraj.simplechess.model.PieceType
+import com.samnagaraj.simplechess.model.Player
 
 class ChessBoardView @JvmOverloads constructor(
     context: Context,
