+++ app/src/main/java/com/samnagaraj/simplechess/ui/theme/Color.kt
@@ -9,3 +9,6 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+val LightBrown = Color(0xFFE0BFA5)
+val DarkBrown = Color(0xFF8E5C43)
