+++ app/src/main/res/drawable/ic_chess_knight_white.xml
@@ -26,13 +26,19 @@
         android:strokeLineCap="round"
         android:strokeLineJoin="round"
         android:pathData="M 9.5 25.5 A 0.5 0.5 0 1 1 8.5,25.5 A 0.5 0.5 0 1 1 9.5 25.5 z"/>
-    <path
-        android:fillColor="#000000"
-        android:strokeColor="#000000"
-        android:strokeWidth="1.5"
-        android:strokeLineCap="round"
-        android:strokeLineJoin="round"
-        android:pathData="M 15 15.5 A 0.5 1.5 0 1 1  14,15.5 A 0.5 1.5 0 1 1  15 15.5 z"
-        android:transformMatrix="0.866,0.5,-0.5,0.866,9.693,-5.173"/>
+    <group
+        android:scaleX="0.866"
+        android:scaleY="0.866"
+        android:translateX="9.693"
+        android:translateY="-5.173"
+        android:rotation="30">
+      <path
+          android:fillColor="#000000"
+          android:strokeColor="#000000"
+          android:strokeWidth="1.5"
+          android:strokeLineCap="round"
+          android:strokeLineJoin="round"
+          android:pathData="M 15 15.5 A 0.5 1.5 0 1 1  14,15.5 A 0.5 1.5 0 1 1  15 15.5 z"/>
+    </group>
   </group>
 </vector>
